# reference/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LearnPath - Personalized Learning & Implementation Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /saved-cards:
    get:
      operationId: listSavedCards
      tags: [saved-cards]
      summary: Get user's saved summary cards
      responses:
        "200":
          description: List of saved cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedCard"

    post:
      operationId: saveCard
      tags: [saved-cards]
      summary: Save a summary card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveCardInput"
      responses:
        "201":
          description: Card saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedCard"
  /saved-cards/{cardId}:
    delete:
      operationId: deleteSavedCard
      tags: [saved-cards]
      summary: Remove a saved card
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Card deleted

  # ROADMAPS
...
components:
  schemas:
...
    SaveCardInput:
      type: object
      required: [contentId, lessonId]
      properties:
        contentId:
          type: integer
        lessonId:
          type: integer
    SavedCard:
      type: object
      required: [id, contentId, lessonId, contentTitle, lessonTitle, keyTakeaways, coverColor, savedAt]
      properties:
        id:
          type: integer
        contentId:
          type: integer
        lessonId:
          type: integer
        contentTitle:
          type: string
        lessonTitle:
          type: string
        keyTakeaways:
          type: array
          items:
            type: string
        coreInsight:
          type: ["string", "null"]
        coverColor:
          type: string
        accentColor:
          type: ["string", "null"]
        savedAt:
          type: string
          format: date-time

    # ROADMAPS
...